pyproject.toml
[project]
name = "terminal-bench-dummy-purple"
version = "0.1.0"
description = "Dummy purple agent for terminal-bench evaluations"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk[http-server]>=0.3.20",
    "brotli>=1.1.0",
    "httpx>=0.28.1",
    "uvicorn>=0.38.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.1",
]
